C2Update11

+++ tenmo-client/src/main/java/com/techelevator/tenmo/model/Account.java
@@ -4,15 +4,15 @@ import java.math.BigDecimal;
 
 public class Account {
 
-    private Long accountId;
+    private int accountId;
     private int userId;
     private BigDecimal balance;
 
-    public Long getAccountId() {
+    public int getAccountId() {
         return accountId;
     }
 
-    public void setAccountId(Long accountId) {
+    public void setAccountId(int accountId) {
         this.accountId = accountId;
     }
